fix: returns correct webpack execution code status + initialize args.json only if file not exist

+++ webdev/javascript/webdev.sh
@@ -69,6 +69,8 @@ case "$1" in
         update_deploy_script;;
 
     build )
-        npx --package=webpack@5.72.0 -- webpack --config ./azion/webpack.config.js -o ${OUTPUT_DIR} --mode production
-        echo "{}" > ./args.json;;
+        if [ ! -f ./args.json ]; then
+            echo "{}" > ./args.json
+        fi
+        npx --package=webpack@5.72.0 -- webpack --config ./azion/webpack.config.js -o ${OUTPUT_DIR} --mode production || exit $? ;;
 esac
